% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accept_curve.R
\name{accept_curve}
\alias{accept_curve}
\title{(Sub-)population acceptability curves.}
\usage{
accept_curve(
  forest,
  from,
  to,
  length.out,
  subset = NULL,
  compare = FALSE,
  robust.se = FALSE,
  labels = NULL
)
}
\arguments{
\item{forest}{A trained CEA forest.}

\item{from}{Lowest WTP to consider.}

\item{to}{Highest WTP to consider.}

\item{length.out}{Number of increments from lowest to highest WTP.}

\item{subset}{A specified subpopulation. See grf::average_treatment_effect for details.}

\item{compare}{A logical statement. If TRUE, the subpopulation curve is compared to the sample average curve. Defaults to FALSE.}

\item{robust.se}{Whether or not robust (sandwich) standard errors should be used to compute the acceptance probability. Defaults to FALSE.}
}
\value{
A matrix with acceptance probabilities given the data, subpopulation and different choices of WTP.
}
\description{
\code{accept_curve} Provides acceptability curves for the entire sample or for a specified subpopulation.
}
\examples{
\dontrun{
To be added...
}
}
