% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_effects.R
\name{avg_effects}
\alias{avg_effects}
\title{Estimate average effects after CEA forests.}
\usage{
avg_effects(
  forest,
  WTP = NULL,
  subset = NULL,
  robust.se = FALSE,
  ci.level = 0.95,
  compliance.scores = NULL,
  icer.ci = TRUE,
  boot.ci = FALSE,
  R = 999
)
}
\arguments{
\item{forest}{A trained CEA forest.}

\item{WTP}{Willingness to pay per one-unit increase in Y.}

\item{subset}{A specified subpopulation to obtain average effects for (optional).}

\item{robust.se}{Whether or not robust (sandwich) standard errors are desired. Defaults to FALSE. Bootstrapped CIs or ICER CIs are not affected by this setting.}

\item{ci.level}{The desired confidence level.}

\item{compliance.scores}{An optional two-column matrix containing pre-fitted compliance scores for instrumental forests (col 1 = outcomes, col 2 = costs).}

\item{icer.ci}{Logical, if confidence intervals for ICERs are desired. Uses Fieller's method if boot.ci=FALSE. Defaults to TRUE.}

\item{boot.ci}{Logical, if bootstrapped confidence intervals (BCa) are desired. Defaults to FALSE.}

\item{R}{Number of bootstrap replicates for bootstrapped CIs. Defaults to 999.}
}
\value{
Returns de-biased, pooled effect estimates with asymptotic variance estimates or accelerated percentile bootstrap confidence intervals if boot.ci=TRUE. Confidence intervals for ICERs are estimated using Fieller's method unless boot.ci=TRUE.
}
\description{
\code{avg_effects} Estimate de-biased average causal effects for the entire sample or for a specified subpopulation.
}
\examples{
\dontrun{
To be added...
}
}
\references{
Chernozhukov, V., Chetverikov, D., Demirer, M., Duflo, E., Hansen, C., & Newey, W. (2017). Double/debiased/neyman machine learning of treatment effects. American Economic Review, 107(5), 261-65.
}
