% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_effects.R
\name{boot.dr_scores}
\alias{boot.dr_scores}
\title{Bootstrap average effects}
\usage{
boot.dr_scores(dc, dy, R, WTP, alpha)
}
\arguments{
\item{dc}{delta_cost}

\item{dy}{delta_y}

\item{alpha}{Desired confidence level.}
}
\value{
Returns a matrix with estimated standard errors and BCa confidence intervals.
}
\description{
\code{boot.dr_scores} Bootstraps doubly robust scores and obtains accelerated bootstrap confidence intervals (BCa).
}
\keyword{internal}
