% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cea_policy_tree.R
\name{boot.policy_scores}
\alias{boot.policy_scores}
\title{Bootstrap average effects}
\usage{
boot.policy_scores(Gamma_all, Gamma_policy, R, alpha)
}
\arguments{
\item{Gamma_all}{Scores for treating everyone vs treating no-one.}

\item{Gamma_policy}{Scores for suggested policy vs treating no-one.}

\item{R}{Number of bootstrap replicates.}

\item{alpha}{Desired confidence level.}
}
\value{
Returns a matrix with estimated standard errors and BCa confidence intervals.
}
\description{
\code{boot.dr_scores} Bootstraps doubly robust scores and obtains accelerated bootstrap confidence intervals (BCa).
}
\keyword{internal}
