% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_plane.R
\name{ce_plane}
\alias{ce_plane}
\title{Plot an individual-level or average cost-effectiveness plane using a CEA forest.}
\usage{
ce_plane(
  forest,
  alpha = 0.05,
  type = "individual",
  R = 999,
  subset = NULL,
  WTP = NULL,
  certainty_groups = FALSE
)
}
\arguments{
\item{forest}{A trained CEA forest.}

\item{alpha}{The desired confidence level, defaults to 0.05.}

\item{type}{A string (either "individual" or "average") specifying which type of cost-effectiveness plane to plot. Defaults to individual.}

\item{R}{Number of bootstrap replicates (only used for average CE planes).}

\item{subset}{A specified subpopulation to compute the plane for (only used for average CE planes). Defaults to NULL.}

\item{WTP}{The willingness to pay per one unit increase in the outcome. Not used when certainty_groups is TRUE. Re-estimate the forest to change WTP in that case. Uses the WTP from the forest object by default.}

\item{certainty_groups}{Color code individual estimates based on certainty groups? Defaults to FALSE.}
}
\value{
A ggplot2 object.
}
\description{
\code{ce_plane} Plots the results from CEA forests, optionally with uncertainty estimates from an NMB forest.
}
\examples{
\dontrun{
To be added...
}
}
