% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cea_policy_tree.R
\name{cea_create_dot_body}
\alias{cea_create_dot_body}
\title{Writes each node information
If it is a leaf node: show it in different color, show number of samples, show leaf id
If it is a non-leaf node: show its splitting variable and splitting value}
\usage{
cea_create_dot_body(
  tree,
  index = 1,
  group.names = c("Control treatment", "New treatment")
)
}
\arguments{
\item{tree}{the tree to convert}

\item{index}{the index of the current node}

\item{group.names}{names of the treatment and control states (defaults to "Control treatment", "New treatment")}
}
\description{
Writes each node information
If it is a leaf node: show it in different color, show number of samples, show leaf id
If it is a non-leaf node: show its splitting variable and splitting value
}
\keyword{internal}
