% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cea_forest.R
\name{cea_forest}
\alias{cea_forest}
\title{Estimate causal forests for outcomes, costs and net monetary benefits.}
\usage{
cea_forest(
  Y,
  C,
  X,
  W,
  Z = NULL,
  WTP = NULL,
  W.hat = NULL,
  tune.parameters = "all",
  num.trees = 5000,
  ...
)
}
\arguments{
\item{Y}{The outcome vector.}

\item{C}{The cost vector.}

\item{X}{The covariate matrix.}

\item{W}{The treatment vector.}

\item{Z}{An instrumental variable. (Optional)}

\item{WTP}{Willingness to pay per one-unit increase in the outcome. Defaults to 1.}

\item{W.hat}{Pre-fitted propensity scores for treatment (W). If NULL, the algorithm fits a regression forest to estimate W.hat.}

\item{tune.parameters}{Which hyperparameters to tune. Defaults to "all". See grf::causal_forest for other options. Option "none" uses default settings for all parameters.}

\item{num.trees}{The number of trees in each forest. Defaults to 5000. Can (and probably should) be set to a higher number to reduce Monte Carlo errors.}

\item{...}{Other options to be passed to grf::causal_forest() or grf::instrumenal_forest() if instrument is supplied.}
}
\value{
Returns a list containing two causal forest objects (one for the outcome and one for costs). If an instrument is supplied, the code returns two instrumental forest objects.
}
\description{
\code{cea_forest} Runs causal forests for outcomes, costs and net monetary benefits given a specified willingness to pay (a wrapper for grf::causal_forest).
}
\examples{
\dontrun{
To be added...
}
}
\references{
Athey, S., Tibshirani, J., & Wager, S. (2019). Generalized random forests. The Annals of Statistics, 47(2), 1148-1178.
}
