% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_forest.R
\name{debias_effects}
\alias{debias_effects}
\title{Get de-biased heterogeneous effect estimates (for custom explainers)}
\usage{
debias_effects(forest, WTP = NULL)
}
\arguments{
\item{forest}{A trained CEAforests object.}

\item{WTP}{The willingness to pay for a one-unit increase in Y. Defaults to the WTP supplied to the CEAforests object unless manually specified.}
}
\value{
Returns a matrix of transformed outcomes that can be used in parametric or semi-parametric regressions to explain the results from the forest.
}
\description{
\code{debias_effects} Provides de-biased tau estimates based on AIPW (for use in custom explainers, e.g., regression models or GAMs).
}
\examples{
\dontrun{
To be added...
}
}
