% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_forest.R
\name{dr.cate.regress}
\alias{dr.cate.regress}
\title{Internal function.}
\usage{
dr.cate.regress(
  dr.scores,
  X = NULL,
  alpha = 0.05,
  robust.se = TRUE,
  subset = NULL
)
}
\arguments{
\item{dr.scores}{Doubly robust scores for tau.}

\item{X}{The variables to include in the model. If NULL, and intercept only model is run.}

\item{alpha}{The desired significance level, defaults to 0.05.}

\item{robust.se}{Whether or not to compute robust (sandwich) standard errors. Defaults to TRUE.}

\item{subset}{A specified subset.}
}
\value{
Returns the results from linear regression on the estimated heterogeneous effects with supplied covariates with asymptotically valid standard errors. To be used with causal_forest object directly. For cea_forests or nmb_forests, use the explain_forest() function.
}
\description{
\code{dr.cate.regress} Fits a linear regression to the estimated heterogeneous effects to assess the determinants of heterogeneity. Uses augmented inverse probability weighting (AIPW) to de-bias the scores, with sandwich standard errors.
}
\examples{
\dontrun{
To be added...
}
}
\keyword{internal}
