% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_forest.R
\name{explain_forest}
\alias{explain_forest}
\title{Explain the results from CEA forests using best linear projection.}
\usage{
explain_forest(
  forest,
  X = NULL,
  alpha = 0.05,
  WTP = NULL,
  robust.se = TRUE,
  subset = NULL
)
}
\arguments{
\item{forest}{A trained CEA forest.}

\item{X}{The variables to include in the model. If NULL, and intercept only model is run, which gives the ATE.}

\item{alpha}{The desired significance level, defaults to 0.05.}

\item{WTP}{The WTP for the net monetary benefit forest. Uses the WTP supplied to the CEA forest if NULL.}

\item{robust.se}{If robust (sandwich) standard errors are desired. Defaults to TRUE.}

\item{subset}{A specified subset of the data to compute the regression on.}
}
\value{
Returns the results from linear regression(s) on the estimated heterogeneous effects with supplied covariates with asymptotically valid standard errors.
}
\description{
\code{explain_forest} Fits a linear regression to the estimated heterogeneous effects to assess the determinants of heterogeneity. Uses augmented inverse probability weighting (AIPW) to de-bias the scores.
}
\examples{
\dontrun{
To be added...
}
}
\references{
Chernozhukov, Victor, and Vira Semenova. "Simultaneous inference for Best Linear Predictor of the Conditional Average Treatment Effect and other structural functions." arXiv preprint arXiv:1702.06240 (2017).
}
