% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_effects.R
\name{icer.ci}
\alias{icer.ci}
\title{ICERs with Fieller's method CIs}
\usage{
icer.ci(dc, dy, alpha)
}
\arguments{
\item{dc}{delta_cost}

\item{dy}{delta_y}

\item{alpha}{Desired confidence level.}
}
\value{
Returns ICER with confidence intervals based on Fieller's method assuming a bivariate normal distribution for the averages of delta costs and delta y.
}
\description{
\code{fieller} Fieller's method confidence intervals for ICERs.
}
\keyword{internal}
