% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cea_policy_tree.R
\name{infer_policy}
\alias{infer_policy}
\title{Conduct inference for a personalized treatment policy.}
\usage{
infer_policy(
  forest,
  treat.policy,
  WTP = NULL,
  ci.level = 0.95,
  robust.se = FALSE,
  boot.ci = FALSE,
  R = 999
)
}
\arguments{
\item{forest}{A trained CEA forest.}

\item{treat.policy}{A logical vector or cea policy tree defining the subset covered by the policy.}

\item{WTP}{Willingness to pay for a one unit increase in the outcome. If NULL, the WTP supplied to the CEA forest is used.}

\item{ci.level}{Desired significance level (for confidence intervals).}

\item{robust.se}{Whether or not robust (sandwich) standard errors are desired. Defaults to FALSE. Ignored when boot.ci=TRUE.}

\item{boot.ci}{Whether or not bootstrapped confidence intervals are desired. Defaults to FALSE.}

\item{R}{The number of bootstrap replications. Defaults to 999. Ignored when boot.ci=FALSE.}
}
\value{
Returns a matrix containing estimates for the average welfare gain per population member under various treatment policies (treat everyone vs. treat no one; treat suggested subset vs. treat no one; treat suggested subset vs. treat everyone). Also outputs the share of the popuation covered by the policy.
}
\description{
Conduct inference for a personalized treatment policy, either using a manually specified policy or a learned policy.
}
\examples{
\dontrun{
To be added...
}
}
