% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cea_policy_tree.R
\name{plot.cea_policy_tree}
\alias{plot.cea_policy_tree}
\title{Plot a cea_policy_tree tree object.}
\usage{
\method{plot}{cea_policy_tree}(x, group.names = c("Control treatment", "New treatment"), ...)
}
\arguments{
\item{x}{The tree to plot}

\item{group.names}{names of the treatment and control states (defaults to "Control treatment", "New treatment")}

\item{...}{Additional options (currently ignored).}
}
\description{
Plot a cea_policy_tree tree object.
}
