% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cea_forest.R
\name{plot.CEAforests}
\alias{plot.CEAforests}
\title{Plotting function for CEA forests.}
\usage{
\method{plot}{CEAforests}(
  forest,
  which.y = "all",
  which.x = NULL,
  conditional = FALSE,
  smooth = FALSE,
  ci.level = 0.95,
  x.range = NULL,
  length.out = 100,
  xlab = "X",
  labels = "AUTO",
  ...
)
}
\arguments{
\item{forest}{A trained CEA forest.}

\item{which.y}{A string or string vector naming which outcomes to plot (any combination of "outcome", "cost" and "nmb" is acceptable). Defaults to "all", which produces three plots in one-row, three-column grid.}

\item{which.x}{A column number or string naming a single variable from the X matrix in the CEAforests object. If null, the function outputs histograms of the out-of-bag estimates.}

\item{conditional}{Whether or not to keep all other variables in the X matrix constant at their mean in bivariate plots (defaults to FALSE). Ignored if which.x is null.}

\item{smooth}{Whether or not to plot a semi-parametric smooth function fit to doubly robust scores for tau(x) (via the mgcv package) instead of out-of-bag estimates (unconditional) or non-parametric predictions (conditional). Defaults to FALSE.}

\item{ci.level}{The desired confidence level for confidence intervals (used when applicable). Defaults to 0.95.}

\item{x.range}{A two-element numeric vector that controls the range of the x-axis. Defaults to min(which.x) and max(which.x).}

\item{length.out}{The length of the sequence of X values to be plotted (used when conditional=TRUE). Defaults to 100.}

\item{xlab}{Label for the X axis. Defaults to the column name for the focal X variable, or to "X" if no column name is available.}

\item{labels}{Labels to be passed to cowplot::plot_grid when more than one outcome type (which.y) is supplied. Defaults to "AUTO", which labels the plots using letters.}

\item{...}{Additional arguments to be passed to the gam function in mgcv. Ignored if smooth=FALSE.}
}
\value{
A ggplot via ggplot2 or grid of ggplots via the cowplot package.
}
\description{
Provides histograms, scatter plots or (partial) effects plots to assess heterogeneity with respect to a covariate after a CEAforest.
}
\examples{
\dontrun{
To be added...
}
}
