% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_plane.R
\name{ce_plane}
\alias{ce_plane}
\title{Plot an individual or population-level cost-effectiveness plane using a CEA forest.}
\usage{
ce_plane(
  forest,
  alpha = 0.05,
  type = "individual",
  R = 999,
  subset = NULL,
  WTP = NULL
)
}
\arguments{
\item{forest}{A trained CEA forest.}

\item{alpha}{The desired confidence level, defaults to 0.05. Only used when nmb is supplied.}

\item{type}{A string (either "individual" or "population") specifying which type of cost-effectiveness plane to plot. Defaults to individual.}

\item{R}{Number of bootstrap replicates (only used for population-level planes).}

\item{subset}{A specified subpopulation to compute the plane for (only used for population-level planes). Defaults to NULL.}

\item{WTP}{The willingness to pay per one unit increase in the outcome. Only used for population-level plots. Re-estimate the forest to change WTP for individual-level plots. Uses the WTP from the forest object by default.}
}
\value{
A ggplot2 object.
}
\description{
\code{ce_plane} Plots the results from CEA forests, optionally with uncertainty estimates from an NMB forest.
}
\examples{
\dontrun{
To be added...
}
}
