% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cea_policy_tree.R
\name{cea_export_graphviz}
\alias{cea_export_graphviz}
\title{Export a tree in DOT format.
This function generates a GraphViz representation of the tree,
which is then written into `dot_string`.}
\usage{
cea_export_graphviz(tree, group.names = c("Do not reimburse", "Reimburse"))
}
\arguments{
\item{tree}{the tree to convert}

\item{group.names}{names of the treatment and control states (defaults to c("Do not reimburse", "Reimburse"))}
}
\description{
Export a tree in DOT format.
This function generates a GraphViz representation of the tree,
which is then written into `dot_string`.
}
\keyword{internal}
