% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accept_curve.R
\name{plot.accept_curve}
\alias{plot.accept_curve}
\title{Plotting function for (sub)-population acceptability curves.}
\usage{
\method{plot}{accept_curve}(x, labels = NULL, ...)
}
\arguments{
\item{x}{An acceptability curve object.}

\item{labels}{A character vector of labels for the selected subset, the entire sample and the unselected subset, defaults to c("Selected subgroup", "Full sample", "Not in selected subgroup").}

\item{...}{Additional options (currently ignored).}
}
\value{
A ggplot.
}
\description{
\code{plot.accept_curve} Plots acceptability curves using ggplot2.
}
\examples{
\dontrun{
To be added...
}
}
\keyword{internal}
